Utilizing getDaysLeft to display the amount of days left for a particular campaign
