+++ lib/index.ts
@@ -16,3 +16,4 @@ export { callApi } from "./helpers/callApi";
 export { cn } from "./helpers/cn";
 export { parseJSON } from "./helpers/parseJSON";
 export { omitKeys } from "./helpers/omitKeys";
+export { getDaysLeft } from "./helpers/getDaysLeft";
